requirements.txt
aiohttp==3.9.0
aiofiles==23.2.0
python-dotenv==0.21.0